@@ -4,12 +4,12 @@ apply plugin: 'com.github.dcendents.android-maven'
 group = 'com.github.zj565061763'
 
 android {
-    compileSdkVersion 27
+    compileSdkVersion 28
     defaultConfig {
         minSdkVersion 15
         targetSdkVersion 22
-        versionCode 20190601
-        versionName "1.1.1"
+        versionCode 20190723
+        versionName "1.1.2"
     }
 
     buildTypes {
@@ -21,7 +21,7 @@ android {
 }
 
 dependencies {
-    compileOnly 'com.android.support:support-v4:+'
+    implementation 'androidx.appcompat:appcompat:1.0.2'
 }
 
 // 指定编码
