@@ -1,8 +1,9 @@
 package com.sd.lib.viewpager.helper;
 
 import android.database.DataSetObserver;
-import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
+
+import androidx.viewpager.widget.PagerAdapter;
+import androidx.viewpager.widget.ViewPager;
 
 import java.lang.ref.WeakReference;
 
