@@ -1,14 +1,15 @@
 package com.sd.viewpager_helper;
 
 import android.os.Bundle;
-import android.support.v4.view.PagerAdapter;
-import android.support.v4.view.ViewPager;
-import android.support.v7.app.AppCompatActivity;
 import android.util.Log;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.LinearLayout;
+
+import androidx.appcompat.app.AppCompatActivity;
+import androidx.viewpager.widget.PagerAdapter;
+import androidx.viewpager.widget.ViewPager;
 
 import com.sd.lib.viewpager.helper.FPagerPercentChangeListener;
 import com.sd.lib.viewpager.helper.FPagerSelectedChangeListener;
